% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Takes in a trained model with X and y values to produce a confusion matrix
visual. If predicted_y array is passed in,other evaluation scoring metrics such as
Recall, and precision will also be produced}
\usage{
confusion_matrix(actual_y, predicted_y, labels = NULL, title = NULL)
}
\arguments{
\item{actual_y}{a data frame of the validation labels}

\item{predicted_y}{a data frame of predicted values with model}

\item{labels}{a list of string of confusion matrix labels}

\item{title}{a string of the confusion matrix label}
}
\value{
A plot and a dataframe with scoring metrics
}
\description{
Implementation based on https://stackoverflow.com/a/59119854
}
\examples{
confusion_matrix(as.factor(c("yes", "no", "yes")), as.factor(c("no", "no", "yes")))
}
