% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_table.R
\name{model_comparison_table}
\alias{model_comparison_table}
\title{model_comparison_table
Takes in training data, testing data,
with the target as the last column and
fitted models with meaningful names,
then generates a model comparison
table.}
\usage{
model_comparison_table(train_data, test_data, ...)
}
\arguments{
\item{train_data}{tibble of training data with target as last column}

\item{test_data}{tibble of testing data with target as last column}

\item{...}{fitted models assigned to meaningful model names}
}
\value{
tibble of results, allowing you to compare models
}
\description{
model_comparison_table
Takes in training data, testing data,
with the target as the last column and
fitted models with meaningful names,
then generates a model comparison
table.
}
\examples{
\dontrun{
model_comparison_table(train_set, test_set,
model1 = model_lm, model2 = model_gbm)
}
}
