% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomptab.R
\name{modelcomptab}
\alias{modelcomptab}
\title{Title}
\usage{
modelcomptab(models, X_train, y_train, X_test, y_test, scoring = "default")
}
\arguments{
\item{models}{list}

\item{X_train}{tibble}

\item{y_train}{vector}

\item{X_test}{tibble}

\item{y_test}{vector}

\item{scoring}{str}
}
\value{

}
\description{
Title
}
\examples{
modelcomptab(models, X_train, y_train, X_test, y_test)
}
