% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{creates a ROC plot}
\usage{
plot_roc(y_label, predict_prob)
}
\arguments{
\item{y_label}{a vector contains the validation set label}

\item{predict_prob}{a vector contains the prediction probabilities}
}
\value{
A plot
}
\description{
Takes in a vector of prediction labels and a vector of prediction probabilities.
}
\examples{
plot_roc(df$labels,df$predictions)
}
