% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Takes in a trained model with X and y values to produce a confusion matrix
visual. If predicted_y array is passed in,other evaluation scoring metrics such as
Recall, and precision will also be produced}
\usage{
confusion_matrix(model, X_train, X_test, y_train, y_test,
  predicted_y = NULL, labels = NULL, title = NULL)
}
\arguments{
\item{model}{trained model}

\item{X_train}{a data frame of the training dataset without labels}

\item{X_test}{a data frame of the validation dataset without labels}

\item{y_train}{a data frame of the training labels}

\item{y_test}{a data frame of the validation labels}

\item{predicted_y}{a data frame of predicted values with model}

\item{labels}{a list of string of confusion matrix labels}

\item{title}{a string of the confusion matrix label}
}
\value{
A plot
}
\description{
Takes in a trained model with X and y values to produce a confusion matrix
visual. If predicted_y array is passed in,other evaluation scoring metrics such as
Recall, and precision will also be produced
}
\examples{
confusion_matrix(svm, X_train, y_train, X_valid, y_valid)
}
