% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{creates a ROC plot}
\usage{
plot_roc(y_label, predict_proba)
}
\arguments{
\item{y_label}{a vector contains the validation set label}

\item{predict_proba}{a vector contains the prediction probabilities}
}
\value{
A plot
}
\description{
Takes in a vector of prediction labels and a vector of prediction probabilities.
}
\examples{
plot_roc(c(1,0,0,1,1), c(0.5,0.34,0.89,0.12,0.99))
}
