% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_train_valid_acc.R
\name{plot_train_valid_acc}
\alias{plot_train_valid_acc}
\title{Plot training and validation dataset accuracies}
\usage{
plot_train_valid_acc(model_name, train, valid, param_name, param_vec)
}
\arguments{
\item{model_name}{a string of the machine learning model name}

\item{train}{a data frame of the training data}

\item{valid}{a data frame of the validation data}

\item{param_name}{a string of the parameter name}

\item{param_vec}{a vector of the parameter values}
}
\value{
A plot
}
\description{
Plot training and validation dataset accuracies
}
\examples{
plot_train_valid_acc(iris$Species)
}
